requirements.txt
streamlit
beautifulsoup4
playwright
openai
python-dotenv
tqdm
pinecone-client
